+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Prompt customization configuration model for personalizing command prompts
+
 ## [0.10.0] - 2026-01-24
 
 ### Added